Suppress `clippy::needless_maybe_sized` for `serde`.
This change `expect`s the `clippy::needless_maybe_sized` lint when the
`serde` feature is enabled. The implementations of `serde` traits for
`Text<T>` require that `T: Clone` and therefore `T: Sized`. This ignores
the `T: ?Sized` bound on `Text<T>`, but this is appropriate and
intentional. The `T: ?Sized` bound is necessary for certain type
constructions, namely `TStr`.